+++ MIE/SSEClient.cpp
@@ -351,17 +351,17 @@ set<QueryResult,cmp_QueryResult> SSEClient::calculateQueryResults(int sockfd, ma
     
     //calculate img query results
     timespec start = getTime();                    //start index time benchmark
-    map<int,float>* imgQueryResults = new map<int,float>;
+    map<int,double>* imgQueryResults = new map<int,double>;
 //    for (map<int,map<int,int> >::iterator it=imgPostingLists.begin(); it!=imgPostingLists.end(); ++it) {
     for (int i = 0; i < imgPostingLists.size(); i++) {
 //        float idf = getIdf(DOCS, it->size());
         if (imgPostingLists[i].size() > 0) {
-            const float idf = getIdf(DOCS, imgPostingLists[i].size());
+            const double idf = getIdf(DOCS, imgPostingLists[i].size());
 //          for (map<int,int>::iterator it2 = it->begin(); it2!=it->end(); ++it2) {
             for (map<int,int>::iterator it2 = imgPostingLists[i].begin(); it2!=imgPostingLists[i].end(); ++it2) {
 //              const int queryTf = (*vws)[it->first];
                 const int queryTf = (*vws)[i];
-                const float score = queryTf * getTfIdf(it2->second, idf);
+                const double score = queryTf * getTfIdf(it2->second, idf);
                 if (imgQueryResults->count(it2->first) == 0)
                     (*imgQueryResults)[it2->first] = score;
                 else
@@ -373,12 +373,12 @@ set<QueryResult,cmp_QueryResult> SSEClient::calculateQueryResults(int sockfd, ma
     free(imgQueryResults);
     
     //calculate text query results
-    map<int,float>* textQueryResults = new map<int,float>;
+    map<int,double>* textQueryResults = new map<int,double>;
     for (map<vector<unsigned char>,map<int,int> >::iterator it=textPostingLists.begin(); it!=textPostingLists.end(); ++it) {
-        float idf = getIdf(DOCS, it->second.size());
+        double idf = getIdf(DOCS, it->second.size());
         for (map<int,int>::iterator it2 = it->second.begin(); it2!=it->second.end(); ++it2) {
             const int queryTf = (*encKeywords)[it->first];
-            const float score = queryTf * getTfIdf(it2->second, idf);
+            const double score = queryTf * getTfIdf(it2->second, idf);
             if (textQueryResults->count(it2->first) == 0)
                 (*textQueryResults)[it2->first] = score;
             else
@@ -437,9 +437,9 @@ set<QueryResult,cmp_QueryResult> SSEClient::mergeSearchResults(set<QueryResult,c
                 rank->second.imgRank = i++;
         }
     }
-    map<int,float> queryResults;
+    map<int,double> queryResults;
     for (map<int,Rank>::iterator it=ranks.begin(); it!=ranks.end(); ++it) {
-        float score = 0.f, df = 0.f;
+        double score = 0.f, df = 0.f;
         if (it->second.textRank > 0) {
             score =  1 / pow(it->second.textRank,2);
             df++;
